Don't overwrite existing namelist for WPS and WRF

+++ components/scripts/common/run_wps.ksh
@@ -51,7 +51,10 @@ ln -sf $WRF_BUILD/WPS-${WPS_VERSION}/*.exe .
 # The Vtable is dependent on the data that is used
 # Will need to pull this in dynamically somehow, tie to data/namelist
 
-cp -f $CASE_DIR/namelist.wps .
+if [[ ! -e namelist.wps ]]; then
+  cp -f $CASE_DIR/namelist.wps .
+fi
+
 cp -f $CASE_DIR/Vtable.GFS Vtable
 
 # Link input data
